requirements.txt
streamlit
python-dotenv
langchain-openai  # Or the actual package name you're using for LangChain